# Added cache for cigar string parsing