Significantly improved performance. Benchmark of CIGAR parsing alone on a real SHOGUN / Bowtie2 alignment file, which has 1342487 alignments:

- Before: 714 ms ± 18.4 ms per loop
- After: 85.3 ms ± 1.51 ms per loop

Benchmark of the entire `plain_mapper` function on 100000 SAM alignments:

- Before: 262 ms ± 2.85 ms per loop
- After: 214 ms ± 1.26 ms per loop

Also see issue #38